renamed insertion_manager and insertion_filters packages
