+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/main/java/ch/ubique/notifyme/sdk/backend/ws/config/WSBaseConfig.java
@@ -25,11 +25,11 @@ import ch.ubique.notifyme.sdk.backend.ws.controller.NotifyMeControllerV2;
 import ch.ubique.notifyme.sdk.backend.ws.controller.NotifyMeControllerV3;
 import ch.ubique.notifyme.sdk.backend.ws.controller.web.WebController;
 import ch.ubique.notifyme.sdk.backend.ws.controller.web.WebCriticalEventController;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.InsertManager;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.BeforeOnsetFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.FakeRequestFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.IntervalThresholdFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.OverlappingIntervalsFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.InsertManager;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.BeforeOnsetFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.FakeRequestFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.IntervalThresholdFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.OverlappingIntervalsFilter;
 import ch.ubique.notifyme.sdk.backend.ws.security.NotifyMeJwtRequestValidator;
 import ch.ubique.notifyme.sdk.backend.ws.security.RequestValidator;
 import ch.ubique.notifyme.sdk.backend.ws.service.PhoneHeartbeatSilentPush;
