+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/main/java/ch/ubique/notifyme/sdk/backend/ws/insertmanager/OSType.java
@@ -1,4 +1,4 @@
-package ch.ubique.notifyme.sdk.backend.ws.insert_manager;
+package ch.ubique.notifyme.sdk.backend.ws.insertmanager;
 
 public enum OSType {
     ANDROID,
