+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/test/java/ch/ubique/notifyme/sdk/backend/ws/insertmanager/InsertManagerTest.java
@@ -1,11 +1,11 @@
-package ch.ubique.notifyme.sdk.backend.ws.insert_manager;
+package ch.ubique.notifyme.sdk.backend.ws.insertmanager;
 
 import ch.ubique.notifyme.sdk.backend.data.NotifyMeDataServiceV3;
 import ch.ubique.notifyme.sdk.backend.model.UserUploadPayloadOuterClass.UploadVenueInfo;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.FakeRequestFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.IntervalThresholdFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.OverlappingIntervalsFilter;
-import ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters.UploadInsertionFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.FakeRequestFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.IntervalThresholdFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.OverlappingIntervalsFilter;
+import ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters.UploadInsertionFilter;
 import ch.ubique.notifyme.sdk.backend.ws.semver.Version;
 import ch.ubique.notifyme.sdk.backend.ws.util.CryptoWrapper;
 import ch.ubique.notifyme.sdk.backend.ws.util.TokenHelper;
