+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/test/java/ch/ubique/notifyme/sdk/backend/ws/insertmanager/insertfilters/OverlappingIntervalsFilterTest.java
@@ -1,60 +1,51 @@
-package ch.ubique.notifyme.sdk.backend.ws.insert_manager.insertion_filters;
+package ch.ubique.notifyme.sdk.backend.ws.insertmanager.insertfilters;
 
 import ch.ubique.notifyme.sdk.backend.model.UserUploadPayloadOuterClass.UploadVenueInfo;
 import com.google.protobuf.ByteString;
-import org.springframework.security.oauth2.jwt.Jwt;
 
 import java.nio.charset.StandardCharsets;
-import java.time.*;
-import java.time.temporal.ChronoUnit;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
-public class BeforeOnsetFilterTest extends UploadInsertionFilterTest {
-
-    // We don't care about the current time in the filter, we just need a common timestamp for all methods below
-    private static final LocalDateTime currentTime = LocalDateTime.of(LocalDate.now(), LocalTime.MIDNIGHT);
-
-
+public class OverlappingIntervalsFilterTest extends UploadInsertionFilterTest {
     @Override
     List<UploadVenueInfo> getValidVenueInfo() {
-        final List<UploadVenueInfo> venueInfoList = new ArrayList<>();
-        LocalDateTime start = currentTime.minusDays(1);
+        final var venueInfoList = new ArrayList<UploadVenueInfo>();
+        LocalDateTime start = LocalDateTime.now();
         LocalDateTime end = start.plusHours(1);
-        // venue visit one day after onset
-        final var venueInfoCase1 = getVenueInfo(start, end);
-        venueInfoList.add(venueInfoCase1);
-        start = currentTime.minusDays(2);
+        final var venueInfo1 = getVenueInfo(start, end);
+        venueInfoList.add(venueInfo1);
+        start = end.plusMinutes(1);
         end = start.plusHours(1);
-        // venue visit same day as onset
-        final var venueInfoCase2 = getVenueInfo(start, end);
-        venueInfoList.add(venueInfoCase2);
+        final var venueInfo2 = getVenueInfo(start, end);
+        venueInfoList.add(venueInfo2);
         return venueInfoList;
     }
 
     @Override
     List<UploadVenueInfo> getInvalidVenueInfo() {
-        final List<UploadVenueInfo> venueInfoList = new ArrayList<>();
-        LocalDateTime start = currentTime.minusDays(3);
+        final var venueInfoList = new ArrayList<UploadVenueInfo>();
+        final var now = LocalDateTime.now();
+        LocalDateTime start = now;
         LocalDateTime end = start.plusHours(1);
-        // venue visit one day before onset
-        final var venueInfo = getVenueInfo(start, end);
-        venueInfoList.add(venueInfo);
+        final var venueInfo1 = getVenueInfo(start, end);
+        venueInfoList.add(venueInfo1);
+        start = end.minusMinutes(1);
+        end = start.plusHours(1);
+        final var venueInfo2 = getVenueInfo(start, end);
+        venueInfoList.add(venueInfo2);
+        start = now.minusMinutes(59);
+        end = start.plusHours(1);
+        final var venueInfo3 = getVenueInfo(start, end);
+        venueInfoList.add(venueInfo3);
         return venueInfoList;
     }
 
     @Override
     UploadInsertionFilter insertionFilter() {
-        return new BeforeOnsetFilter();
-    }
-
-    @Override
-    public Jwt getToken(LocalDateTime now) throws Exception {
-        final var onset = currentTime.minusDays(2).truncatedTo(ChronoUnit.DAYS).format(DATE_FORMATTER);
-        final var expiry = currentTime.plusMinutes(5).toInstant(ZoneOffset.UTC);
-        return jwtDecoder.decode(tokenHelper.createToken(
-                onset, "0", "notifyMe", "userupload", Date.from(expiry), true, currentTime.toInstant(ZoneOffset.UTC)));
+        return new OverlappingIntervalsFilter();
     }
 
     private UploadVenueInfo getVenueInfo(
@@ -84,6 +75,4 @@ public class BeforeOnsetFilterTest extends UploadInsertionFilterTest {
                 .setFake(false)
                 .build();
     }
-
 }
-
